requirements.txt
transformers
datasets
accelerate
vllm
scikit-learn
pyyaml
numpy
pandas
